refactor: resolved the issue in the namimg of the components and uses
