+++ apps/feedahelp/src/components/ui/features/ingredients/Ingredients.tsx
@@ -5,11 +5,6 @@ const Ingredients = () => {
   return (
     <div className="mb-2 h-96 w-full border border-red-700">
       This is the Ingredients section.
-
-      <div>
-        <p>hello this is for test the generic link: </p> <br />
-        <GenericLink color='green' href="http://www.facebook.com" > Facebook</GenericLink>
-      </div>
     </div>
   );
 };
